requirements.txt
pycryptodome==3.19.0
fastapi>=0.115.7
pydantic>=2.10.4
SQLAlchemy~=2.0.40